fix: userEvent works consistently between providers
